In NQueen.GUI.csproj changed top node to: <Project Sdk="Microsoft.NET.Sdk"> ... </Project>
Used sbyte instead of int whenever possible
In NQueen.Commin made property Id to Id?

+++ NQueen.Common/Interface/ISimulationResults.cs
@@ -4,7 +4,7 @@ namespace NQueen.Common.Interface
 {
     public interface ISimulationResults
     {
-        int BoardSize { get; }
+        sbyte BoardSize { get; }
 
         IEnumerable<Solution> Solutions { get; }
